+++ JsonToMidiConverter/Models/Song/ChordStroke.cs
@@ -6,6 +6,6 @@ namespace JsonToMidiConverter.Models.Song;
 public sealed class ChordStroke : ISerializable
 {
     public StrokeTechnique Technique { get; set; } = StrokeTechnique.None;
-    public byte Duration { get; set; }
-    public byte StartTimeOffset { get; set; }
+    public double Duration { get; set; }
+    public double StartTimeOffset { get; set; }
 }
